project/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-otp==1.6.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyotp==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
uritemplate==4.2.0
